bump to version 0.12
