+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bevy_aseprite"
-version = "0.11.0"
+version = "0.12.0"
 description = "Bevy aseprite loader"
 edition = "2021"
 resolver = "2"
@@ -13,7 +13,7 @@ members = ["derive", "reader"]
 
 [dependencies]
 anyhow = "1.0.43"
-bevy = { version = "0.11.0", default-features = false, features = [
+bevy = { version = "0.12.0", default-features = false, features = [
     "bevy_asset",
     "bevy_render",
     "bevy_sprite"
@@ -22,4 +22,4 @@ bevy_aseprite_derive = { path = "./derive", version = "0.3" }
 bevy_aseprite_reader = { path = "./reader", version = "0.1" }
 
 [dev-dependencies]
-bevy = { version = "0.11.0" }
+bevy = { version = "0.12.0" }
