+++ README.MD
@@ -8,6 +8,7 @@ Compatability table
 
 | bevy | bevy_aseprite |
 | ---- | ------------- |
+| 0.12 | 0.12          |
 | 0.11 | 0.11          |
 | 0.10 | 0.10          |
 | 0.9  | 0.9           |
